requirements.txt
fastapi
uvicorn
pydantic
Pillow
google-generativeai